refactor: Double 타입을 Decimal로 변경

+++ OpenMarket/OpenMarket/Model/Product.swift
@@ -6,9 +6,9 @@ struct Product: Decodable {
     let name: String
     let thumbnail: String
     let currency: Currency
-    let price: Double
-    let bargainPrice: Double
-    let discountedPrice: Double
+    let price: Decimal
+    let bargainPrice: Decimal
+    let discountedPrice: Decimal
     let stock: Int
     let images: [Image]?
     let vendor: Vendor?
@@ -104,7 +104,7 @@ extension Product {
     }
 }
 
-private extension Double {
+private extension Decimal {
     var formatted: String? {
         let numberFormatter = NumberFormatter()
         numberFormatter.numberStyle = .decimal
